% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Binds Two Factors Together}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{factor}

\item{b}{factor}
}
\value{
factor
}
\description{
Binds Two Factors Together
}
\examples{
fbind(iris$Species[c(1,51,101)], PlantGrowth$group[c(1,11,21)])
}

