% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reorder.R
\name{reorder_factor}
\alias{reorder_factor}
\title{Reorder the levels of a factor by frequency}
\usage{
reorder_factor(x)
}
\arguments{
\item{x}{factor}
}
\value{
factor
}
\description{
Reorder the levels of a factor by frequency
}
\examples{
reorder_factor(iris$Species)
}

